refactor(Command): controller 패키지 아래로 이동
