+++ src/main/java/chess/controller/command/End.java
@@ -1,4 +1,4 @@
-package chess.command;
+package chess.controller.command;
 
 import chess.domain.ChessGame;
