+++ src/main/java/chess/controller/command/Status.java
@@ -1,24 +1,23 @@
-package chess.command;
+package chess.controller.command;
 
 import chess.domain.ChessGame;
 import chess.view.OutputView;
 
-public class Start implements Command {
+public class Status implements Command {
 
-    private static final Start INSTANCE = new Start();
+    private static final Status INSTANCE = new Status();
 
-    private Start() {
+    private Status() {
     }
 
-    public static Start getInstance() {
+    public static Status getInstance() {
         return INSTANCE;
     }
 
     @Override
     public void execute(final ChessGame chessGame) {
         try {
-            chessGame.start();
-            OutputView.printChessBoard(chessGame.findAllPiece());
+            OutputView.printStatus(chessGame.calculateScore());
         } catch (final IllegalArgumentException e) {
             OutputView.printError(e.getMessage());
         }
